1.We had added OLED display functionality to visualize the distance readings from the ultrasonic sensor. However, we decided to comment out the OLED-related code in the main.cpp file to focus on testing the RF transmission of distance data without the display. The OLED code can be re-enabled later once we confirm that the RF transmission is working correctly. 2.We had divided the code(oled,RF transmitter,ultrasonic sensor) into separate files for better organization and modularity. 3.But this code did not work as expected on my device(i do not know why)

+++ src/UltrasonicSensor.cpp
@@ -1,4 +1,4 @@
-# include "UltrasonicSensor.h"
+#include "UltrasonicSensor.h"
 
 void UltrasonicSensor::begin()
 {
@@ -6,7 +6,7 @@ void UltrasonicSensor::begin()
     pinMode(echoPin, INPUT);
 }
 
-long UltrasonicSensor::getDistance()
+float UltrasonicSensor::getDistance()
 {
     digitalWrite(trigPin, LOW);
     delayMicroseconds(2);
